Handle per-language conditional required fields.
In settings add LANGUAGES_REPLACEMENT with {'en' : 'Coming soon'} to make english optional even if the field is required.
The core idea is to made available replacement texts that can be outputed by the views in place of the empty fields, depending of the language.
This 'on-the-fly' replacement mechanism is not (maybe never) implemented in this project.